# api/docs/swagger.yaml
definitions:
...
  ek_entity_service.City:
    properties:
      code:
        type: integer
      id:
        type: string
      name:
        type: string
      ru_name:
        type: string
      soato:
        type: integer
    type: object
  ek_entity_service.ConfirmEntityDraftSwag:
    properties:
      comment:
        type: string
      entity_id:
        type: string
      status_id:
        type: string
    required:
    - comment
    - entity_id
    - status_id
    type: object
...
  ek_entity_service.District:
    properties:
      code:
        type: integer
      id:
        type: string
      name:
        type: string
      ru_name:
        type: string
      soato:
        type: integer
    type: object
  ek_entity_service.EntityDraftSwag:
    properties:
      city:
        $ref: '#/definitions/ek_entity_service.City'
        type: object
      comment:
        type: string
      district:
        $ref: '#/definitions/ek_entity_service.District'
        type: object
      entity_gallery:
        items:
          type: string
        type: array
      entity_id:
        type: string
      entity_properties:
        items:
          $ref: '#/definitions/ek_entity_service.EntityProperty'
        type: array
      region:
        $ref: '#/definitions/ek_entity_service.Region'
        type: object
    type: object
...
  ek_entity_service.EntityProperty:
    properties:
      property_id:
        type: string
      value:
        type: string
    type: object
...
  ek_entity_service.Region:
    properties:
      code:
        type: integer
      id:
        type: string
      name:
        type: string
      ru_name:
        type: string
      soato:
        type: integer
    type: object
...
  ek_variables.CreateResponse:
    properties:
      id:
        type: string
    type: object
  ek_variables.EmptyResponse:
    type: object
  ek_variables.FailureResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  ek_variables.SuccessResponse:
    properties:
      success:
        example: true
        type: boolean
    type: object
  entity_service.ApplicantEntity:
    properties:
      name:
        type: string
      phone_number:
        type: string
      user_id:
        type: string
    type: object
...
  entity_service.Entity:
    properties:
      address:
        type: string
      city:
        $ref: '#/definitions/entity_service.EntityCity'
        type: object
      created_at:
        type: string
      district:
        $ref: '#/definitions/entity_service.EntityDistrict'
        type: object
      entity_files:
        items:
          type: string
        type: array
      entity_number:
        type: string
      entity_properties:
        items:
          $ref: '#/definitions/entity_service.GetEntityProperty'
        type: array
      entity_soato:
        type: string
      entity_type_code:
        type: integer
      id:
        type: string
      region:
        $ref: '#/definitions/entity_service.EntityRegion'
        type: object
      registration_date:
        type: string
      status:
        $ref: '#/definitions/entity_service.Status'
        type: object
      updated_at:
        type: string
    type: object
  entity_service.EntityCity:
    properties:
      code:
        type: integer
      id:
        type: string
      name:
        type: string
      ru_name:
        type: string
      soato:
        type: integer
    type: object
  entity_service.EntityDistrict:
    properties:
      code:
        type: integer
      id:
        type: string
      name:
        type: string
      ru_name:
        type: string
      soato:
        type: integer
    type: object
  entity_service.EntityDraft:
    properties:
      applicant:
        $ref: '#/definitions/entity_service.ApplicantEntity'
        type: object
      city:
        $ref: '#/definitions/entity_service.EntityCity'
        type: object
      comment:
        type: string
      created_at:
        type: string
      district:
        $ref: '#/definitions/entity_service.EntityDistrict'
        type: object
      entity:
        $ref: '#/definitions/entity_service.Entity'
        type: object
      entity_draft_number:
        type: string
      entity_draft_soato:
        type: string
      entity_gallery:
        items:
          type: string
        type: array
      entity_id:
        type: string
      entity_properties:
        items:
          $ref: '#/definitions/entity_service.GetEntityProperty'
        type: array
      id:
        type: string
      region:
        $ref: '#/definitions/entity_service.EntityRegion'
        type: object
      status:
        $ref: '#/definitions/entity_service.Status'
        type: object
      updated_at:
        type: string
    type: object
  entity_service.EntityProperty:
    properties:
      property_id:
        type: string
      value:
        type: string
    type: object
  entity_service.EntityRegion:
    properties:
      code:
        type: integer
      id:
        type: string
      name:
        type: string
      ru_name:
        type: string
      soato:
        type: integer
    type: object
...
  entity_service.GetAllEntityDrafts:
    properties:
      applicant:
        $ref: '#/definitions/entity_service.ApplicantEntity'
        type: object
      city:
        $ref: '#/definitions/entity_service.EntityCity'
        type: object
      comment:
        type: string
      created_at:
        type: string
      district:
        $ref: '#/definitions/entity_service.EntityDistrict'
        type: object
      entity:
        $ref: '#/definitions/entity_service.Entity'
        type: object
      entity_draft_number:
        type: string
      entity_draft_soato:
        type: string
      entity_gallery:
        items:
          type: string
        type: array
      entity_id:
        type: string
      entity_properties:
        items:
          $ref: '#/definitions/entity_service.EntityProperty'
        type: array
      id:
        type: string
      region:
        $ref: '#/definitions/entity_service.EntityRegion'
        type: object
      status:
        $ref: '#/definitions/entity_service.Status'
        type: object
      updated_at:
        type: string
    type: object
  entity_service.GetAllEntityDraftsResponse:
    properties:
      count:
        type: integer
      entity_drafts:
        items:
          $ref: '#/definitions/entity_service.GetAllEntityDrafts'
        type: array
    type: object
...
  entity_service.GetEntityProperty:
    properties:
      property:
        $ref: '#/definitions/entity_service.GetProperty'
        type: object
      value:
        type: string
    type: object
...
  entity_service.GetProperty:
    properties:
      collection_name:
        type: string
      description:
        type: string
      id:
        type: string
      is_required:
        type: boolean
      label:
        type: string
      name:
        type: string
      placeholder:
        type: string
      property_options:
        items:
          $ref: '#/definitions/entity_service.PropertyOption'
        type: array
      status:
        type: boolean
      type:
        type: string
      validation:
        type: string
      with_confirmation:
        type: boolean
    type: object
...
  entity_service.PropertyOption:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
  entity_service.Status:
    properties:
      code:
        type: string
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      parent_status_id:
        type: string
      ru_name:
        type: string
      type_code:
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /v1/applicant-entity-draft/{user_id}:
    get:
      consumes:
      - application/json
      description: API for getting all entity drafts by user id
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity_service.GetAllEntityDraftsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      security:
      - ApiKeyAuth: []
      summary: Getting All draft entities by user id
      tags:
      - entity-draft
...
  /v1/entity-draft:
    get:
      consumes:
      - application/json
      description: API for getting all entity drafts
      parameters:
      - in: query
        name: city_id
        type: string
      - in: query
        name: date
        type: string
      - in: query
        name: entity_draft_number
        type: string
      - example: "2021-11-21"
        in: query
        name: from_date
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: region_id
        type: string
      - in: query
        name: status_id
        type: string
      - example: "2021-11-21"
        in: query
        name: to_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity_service.GetAllEntityDraftsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      security:
      - ApiKeyAuth: []
      summary: Getting All draft entities
      tags:
      - entity-draft
    post:
      consumes:
      - application/json
      description: API for creating entity draft
      parameters:
      - description: entity
        in: body
        name: entity
        required: true
        schema:
          $ref: '#/definitions/ek_entity_service.EntityDraftSwag'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ek_variables.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      security:
      - ApiKeyAuth: []
      summary: Create entity draft
      tags:
      - entity-draft
  /v1/entity-draft-expired:
    get:
      consumes:
      - application/json
      description: API for getting expired entity drafts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity_service.GetAllEntityDraftsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Getting expired draft entities
      tags:
      - entity-draft
...
  /v1/entity-draft/{entity_draft_id}:
    delete:
      consumes:
      - application/json
      description: API for deleting entity draft
      parameters:
      - description: entity_draft_id
        in: path
        name: entity_draft_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ek_variables.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Delete entity draft
      tags:
      - entity-draft
    get:
      consumes:
      - application/json
      parameters:
      - description: entity_draft_id
        in: path
        name: entity_draft_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity_service.EntityDraft'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      summary: Get entity draft
      tags:
      - entity-draft
    patch:
      consumes:
      - application/json
      description: API for updating entity draft comment
      parameters:
      - description: entity_draft_id
        in: path
        name: entity_draft_id
        required: true
        type: string
      - description: entity-draft
        in: body
        name: entity-draft
        required: true
        schema:
          $ref: '#/definitions/ek_entity_service.ConfirmEntityDraftSwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ek_variables.EmptyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      security:
      - ApiKeyAuth: []
      summary: Update entity draft comment
      tags:
      - entity-draft
    put:
      consumes:
      - application/json
      description: API for updating entity
      parameters:
      - description: entity_draft_id
        in: path
        name: entity_draft_id
        required: true
        type: string
      - description: entity-draft
        in: body
        name: entity-draft
        required: true
        schema:
          $ref: '#/definitions/ek_entity_service.EntityDraftSwag'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ek_variables.CreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ek_variables.FailureResponse'
      security:
      - ApiKeyAuth: []
      summary: Update entity draft
      tags:
      - entity-draft
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"